requirements.txt
streamlit
langchain
langchain-google-genai
pydantic
python-dotenv
pytest
pytest-mock


